// src/app/interfaces/interfaces.ts
export interface Category {
    id: number;
    name: string;
    image: string;
  }
  
  export interface Product {
    id: number;
    title: string;
    price: number;
    description: string;
    category: Category;
    images: string[];
    quantity?: number; // Nueva propiedad para la cantidad
  }